refactor(back-end): move update threshold & change email routes
