+++ server/src/middleware/check-login.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from 'express'
 import { UNPROTECTED_ROUTES } from '../constants/unprotected-routes'
 
-export default function checkLogin (req: Request, res: Response, next: NextFunction) {
+export function checkLogin (req: Request, res: Response, next: NextFunction) {
   const protectedRoute = UNPROTECTED_ROUTES.findIndex((route) => route === req.path) === -1
 
   if (!req.session!.user && protectedRoute) {
