+++ server/src/middleware/cors.ts
@@ -1,6 +1,6 @@
 import cors from 'cors'
 
-export default function corsHandler () {
+export function corsHandler () {
   return cors({
     origin: 'http://localhost:8080',
     credentials: true
