+++ server/src/routes/user-data/user-data.routes.ts
@@ -4,3 +4,4 @@ import controller from './user-data.controller'
 export default Router()
   .post('/gear', controller.updateGear)
   .post('/spots', controller.updateSpots)
+  .post('/threshold', controller.updateThreshold)
