+++ server/src/routes/user/user.controller.ts
@@ -81,8 +81,33 @@ function logout (req: Request, res: Response) {
   })
 }
 
+async function updateEmail (req: Request, res: Response) {  
+  try {
+    const { email, password }: { email: string, password: string } = req.body
+    const { name }: User = req.session!.user
+
+    await userService.setEmail(name, email, password)
+
+    req.session!.user.email = email
+
+    res.send('OK')
+  } catch (err) {
+    if (err === 401) {
+      res.status(401).send('Incorrect password')
+      return
+    } else if (err === 409) {
+      res.status(409).send('E-mail address is not in the correct format, or already taken')
+      return
+    }
+
+    console.error(err)
+    res.status(500).send()
+  }
+}
+
 export default {
   register,
   login,
-  logout
+  logout,
+  updateEmail
 }
