+++ server/src/routes/user/user.routes.ts
@@ -5,3 +5,4 @@ export default Router()
   .post('/register', controller.register)
   .post('/login', controller.login)
   .post('/logout', controller.logout)
+  .post('/email', controller.updateEmail)
